+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Release Notes
 
+- 0.5.0 *Unreleased*
+    - Add `Stdio.settings` struct for passing rendering settins.
+    - Add emoji support to output as alternative to text.
 - 0.4.0 *2 February 2026*
     - Make `Stdio.ShellCursor` functions Struct functions (ie. static).
     - Rename `Stdio.ShellStyle` and `Stdio.ShellColour` private `tostring()` functions `string()` (non-breaking).